More renaming and import corrections
