+++ test/python/golden/test_ttir_ops_n300.py
@@ -6,8 +6,8 @@ import torch
 import pytest
 
 from typing import List, Tuple
-from ttir_builder.test_utils import compile_to_flatbuffer
-from ttir_builder.ttir_builder import Operand, TTIRBuilder, Shape
+from ttir_builder.utils import compile_to_flatbuffer
+from ttir_builder import Operand, TTIRBuilder, Shape
 
 pytestmark = pytest.mark.n300
