+++ tools/ttir_builder/ccl_golden.py
